fix(test): consolidate MigrationEnvironmentGuard in parent module
- Move MigrationEnvironmentGuard from migration.rs to integration.rs
- Make it pub so all submodules can use it
- Remove duplicate ScopedEnv definition
- Use MigrationEnvironmentGuard::set in mod.rs test

This follows DRY principle and ensures consistent test environment isolation.
